requirements.txt
langchain
langchain_community
faiss-cpu
streamlit
 sentence-transformers
 langchain_openai
 openapi
 PyMuPDF
 reportlab
 python-dotenv
